Added java doc to every method of services. Created custom exceptions

+++ src/main/java/com/wildtac/service/ImageService.java
@@ -1,12 +1,11 @@
 package com.wildtac.service;
 
-import com.wildtac.ImageWasNotFoundException;
 import com.wildtac.domain.Image;
+import com.wildtac.exception.wasnotfound.ImageWasNotFoundException;
 import com.wildtac.repository.ImageRepo;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 @Service
@@ -23,12 +22,12 @@ public class ImageService {
      * The method finds image by field 'id'
      * @param id - field of image
      * @return - found image
-     * @throws EntityNotFoundException - if image was not found by id
+     * @throws ImageWasNotFoundException - if image was not found by id
      */
     public Image getImageById(Long id) {
 
         return imageRepo.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException(String.format("Image with id %s was not found", id)));
+                .orElseThrow(() -> new ImageWasNotFoundException(String.format("Image with id '%s' was not found", id)));
     }
 
     /**
@@ -47,21 +46,36 @@ public class ImageService {
      */
     public Image saveImage(Image image) {
         Image imageCreated = imageRepo.save(image);
-        log.info("New image was saved");
+        log.info(String.format("New image '%s' was saved", imageCreated));
 
         return imageCreated;
     }
 
+    /**
+     * The method deletes image by field 'id'
+     * @param id - field if image
+     * @throws ImageWasNotFoundException - if image was not found by id
+     */
     public void deleteImageById(Long id) {
+        if (!imageRepo.existsById(id)) {
+            throw new ImageWasNotFoundException(String.format("Failed to delete image - image with id '%s' was not found", id));
+        }
         imageRepo.deleteById(id);
     }
 
+    /**
+     * The method find image by it`s parent id and it`s index
+     * @param parentId - id of parent
+     * @param index - the index of image
+     * @return Image - founded image
+     * @throws ImageWasNotFoundException - if image was not found by parentId and index
+     */
     public Image getImageByParentIdAndIndex(Long parentId, Long index) {
 
         return getImagesByParentId(parentId)
                 .stream()
                 .filter(image -> image.getIndex() == index)
                 .findFirst()
-                .orElse(null);
+                .orElseThrow(() -> new ImageWasNotFoundException(String.format("The image with parentId '%s' and index '%s' was not found", parentId, index)));
     }
 }
